Fixed all methods & clean up

+++ Paperless/Paperless.API/Controllers/DocumentController.cs
@@ -7,7 +7,6 @@ using Paperless.DAL.Repositories;
 namespace Paperless.API.Controllers
 {
     //  Ignore BL for now and just query directly to DAL (Sprint 1)
-    //  TODO: implement automapper
     //  TODO: change calls to async
     [ApiController]
     [Route("[controller]")]
@@ -30,11 +29,11 @@ namespace Paperless.API.Controllers
         public IActionResult Get(string id)
         {
             if (!Guid.TryParse(id, out Guid guid))
-                return BadRequest("Invalid Id");
+                return BadRequest("Invalid ID");
         
             DocumentEntity? document = _documentRepository.GetDocumentById(guid);
             if (document == null)
-                return NotFound($"Document {id} not found");
+                return NotFound($"Document ID {id} not found");
         
             return Ok(document);
         }
@@ -42,33 +41,76 @@ namespace Paperless.API.Controllers
          [HttpPost(Name = "PostDocument")] 
          [ProducesResponseType(StatusCodes.Status201Created)]
          [ProducesResponseType(StatusCodes.Status400BadRequest)]
-         public ActionResult<DocumentDTO> Create(DocumentDTO document) {
-            if (document.Id != Guid.Empty) return BadRequest();
-            if (document == null || string.IsNullOrWhiteSpace(document.Name)) {
-                return BadRequest("Invalid document data.");
-            }
+         public ActionResult<DocumentDTO> Create([Bind("Name,Content")] DocumentDTO document) {
+            if (document == null) 
+                return BadRequest("Empty document.");
+
             document.Id = Guid.NewGuid();
+            document.Summary = "summary";
+            document.CreationDate = DateTime.UtcNow;
+            document.Type = "pdf";
+            document.Size = 25;
+
+            if (!CheckDocumentValidity(document))
+                return BadRequest("Invalid document data.");
+
             _documentRepository.InsertDocument(_mapper.Map<DocumentEntity>(document));
+
             return CreatedAtAction(nameof(Get), new { id = document.Id }, document); 
          }
         
         [HttpPut(Name = "PutDocument")]
         public void Put()
         {
-            //  TODO: implement later
+            //  TODO: implement later, edit document
         }
 
         [HttpDelete(Name = "DeleteDocument")]
-        public void DeleteAll()
+        public ActionResult DeleteAll()
         {
-            _documentRepository.DeleteAllDocuments();
+            try
+            {
+                _documentRepository.DeleteAllDocuments();
+                return Ok();
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
         }
 
         [HttpDelete("{id}")]
-        public void Delete(string id)
+        public ActionResult Delete(string id)
         {
-            Guid guid = new Guid(id);
-            _documentRepository.DeleteDocument(guid);
+            try
+            {
+                if (!Guid.TryParse(id, out Guid guid))
+                    return BadRequest("Invalid ID");
+
+                _documentRepository.DeleteDocument(guid);
+                return Ok();
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
+        }
+
+        //  Temporary method
+        private bool CheckDocumentValidity(DocumentDTO document)
+        {
+            if (String.IsNullOrWhiteSpace(document.Name))
+                return false;
+            if (String.IsNullOrWhiteSpace(document.Content))
+                return false;
+            if (String.IsNullOrWhiteSpace(document.Summary))
+                return false;
+            if (String.IsNullOrWhiteSpace(document.Type))
+                return false;
+            if (document.Size <= 0)
+                return false;
+
+            return true;
         }
     }
 }
